+++ Source/WebKit/UIProcess/ios/fullscreen/WKFullScreenWindowControllerIOS.mm
@@ -1593,7 +1593,7 @@ static constexpr NSString *kPrefersFullScreenDimmingKey = @"WebKitPrefersFullScr
     CGPoint translation = [_interactivePanDismissGestureRecognizer translationInView:_fullscreenViewController.get().view];
     CGPoint velocity = [_interactivePanDismissGestureRecognizer velocityInView:_fullscreenViewController.get().view];
     CGFloat progress = translation.y / (_fullscreenViewController.get().view.bounds.size.height / 2);
-    progress = std::min(1., std::max(0., progress));
+    progress = std::min(static_cast<CGFloat>(1.), std::max(static_cast<CGFloat>(0.), progress));
 
     if (_interactivePanDismissGestureRecognizer.get().state == UIGestureRecognizerStateEnded) {
         _inInteractiveDismiss = false;
@@ -1613,7 +1613,7 @@ static constexpr NSString *kPrefersFullScreenDimmingKey = @"WebKitPrefersFullScr
 {
     CGFloat scale = [_interactivePinchDismissGestureRecognizer scale];
     CGFloat velocity = [_interactivePinchDismissGestureRecognizer velocity];
-    CGFloat progress = std::min(1., std::max(0., 1 - scale));
+    CGFloat progress = std::min(static_cast<CGFloat>(1.), std::max(static_cast<CGFloat>(0.), 1 - scale));
 
     CGPoint translation = CGPointZero;
     auto panState = [_interactivePanDismissGestureRecognizer state];
